+++ controllers/k3s/namespace.go
@@ -114,7 +114,7 @@ func (*namespaceController) ReadAll(c *gin.Context) {
 	helper.ResHandler(c, http.StatusOK, &models.Success{
 		Status: "OK",
 		Result: result.Items,
-		Items:  int64(len(result.Items)),
+		Items:  len(result.Items),
 	})
 }
