+++ controllers/k3s/pods.go
@@ -171,6 +171,6 @@ func (*podsController) ReadAll(c *gin.Context) {
 	helper.ResHandler(c, http.StatusOK, &models.Success{
 		Status: "OK",
 		Result: result,
-		Items:  int64(len(result.Items)),
+		Items:  len(result.Items),
 	})
 }
