+++ models/res.go
@@ -5,7 +5,7 @@ type Success struct {
 	Result     interface{} `json:"result" xml:"result"`
 	Page       int         `json:"page,omitempty" xml:"page,omitempty" example:"0"`
 	Limit      int         `json:"limit,omitempty" xml:"limit,omitempty" example:"20"`
-	Items      int64       `json:"items" xml:"items" example:"1"`
+	Items      int         `json:"items" xml:"items" example:"1"`
 	TotalItems int64       `json:"totalItems" xml:"totalItems" example:"20"`
 }
